LayoutTests/platform/gtk/imported/w3c/web-platform-tests/navigation-timing/nav2_test_redirect_xserver-expected.txt

CONSOLE MESSAGE: Error: assert_equals: Expected redirectStart to be 0. expected 0 but got 4
Description

This test validates the values of the window.performance.getEntriesByType("navigation")[0].redirectCount and the window.performance.getEntriesByType("navigation")[0].redirectStart/End times for a cross-origin server side redirect navigation.



FAIL Navigation Timing 2 WPT Error: assert_equals: Expected redirectStart to be 0. expected 0 but got 4
